requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.3.1
numpy>=1.19.5
scikit-learn>=0.24.2
flask
faiss-cpu 